requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic
requests
httpx
pydantic-settings
cachetools
motor 
pymongo
jinja2
itsdangerous
passlib[argon2]
pydantic[email]
python-multipart
PyJWT[crypto]
python-jose[cryptography]
python-dateutil
apscheduler
prometheus-client
openai
webauthn
pywebpush
firebase-admin
